+++ pkgs/dart_mcp/lib/server.dart
@@ -3,4 +3,5 @@
 // BSD-style license that can be found in the LICENSE file.
 
 export 'src/api/api.dart';
+export 'src/server/request_state_codec.dart';
 export 'src/server/server.dart';